requirements.txt
gradio==4.36.1
python-dotenv
google-generativeai
